requirements.txt
scikit-llm
streamlit
python-dotenv
pandas